pyproject.toml
[project]
name = "django-feria-cf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "mysqlclient>=2.2.7",
    "pillow>=11.2.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.2.0",
]
